chat_sync/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
adrf>=0.1,<1.0                    # Async DRF
django-celery-results>=2.5,<3.0

# ASGI and Channels
daphne>=4.0,<5.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0

# Celery
celery>=5.3,<6.0
redis>=5.0,<6.0

# Database
psycopg[binary]>=3.1,<4.0

# Telegram
aiogram>=3.4,<4.0

# DynamoDB (for production)
pynamodb>=6.0,<7.0

# HTTP client
httpx>=0.25,<1.0

# Utilities
python-dateutil>=2.8,<3.0
